add go workspace and refactor to share common pkg
