+++ apps/worker/tasks/userTask.go
@@ -6,8 +6,8 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"job-scheduler/utils"
 	"job-scheduler/worker/models"
-	"job-scheduler/worker/utils"
 	"os"
 	"os/exec"
