--- apps/worker/utils/common.go
@@ -1,12 +1,0 @@
-package utils
-
-import (
-	"errors"
-	"os"
-)
-
-func CheckFileExists(filePath string) bool {
-	_, error := os.Stat(filePath)
-	//return !os.IsNotExist(err)
-	return !errors.Is(error, os.ErrNotExist)
-}
