+++ go.work
@@ -1,0 +1,8 @@
+go 1.22.0
+
+use (
+	./apps/api
+	./apps/scheduler
+	./apps/worker
+	./packages/utils
+)
