+++ packages/utils/go.mod
@@ -1,0 +1,3 @@
+module job-scheduler/utils
+
+go 1.22.0
